Lagt inn test-dependencies i package.json, refaktorert litt i dsop, skrevet snapshot-tester for dsop og 404-side
