+++ package.json
@@ -96,9 +96,12 @@
     "typescript": "^4.8.3"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "12.1.5",
     "@types/calidation": "1.16.1",
     "@types/fetch-mock": "^7.3.5",
     "@types/file-saver": "^2.0.5",
+    "@types/jest": "^29.0.3",
     "@types/js-cookie": "^3.0.2",
     "@types/jsdom": "^20.0.0",
     "@types/node": "18.7.21",
@@ -108,11 +111,13 @@
     "@types/react-modal": "^3.13.1",
     "@types/react-router-dom": "^5.3.3",
     "@types/react-select": "^3.0.19",
+    "@types/react-test-renderer": "17.0.1",
     "classnames": "^2.3.2",
     "less": "^4.1.3",
     "less-plugin-npm-import": "^2.1.0",
     "less-watch-compiler": "^1.16.3",
     "prettier": "^2.7.1",
+    "react-test-renderer": "17.0.2",
     "ts-jest": "^29.0.2",
     "tslint": "^6.1.3",
     "tslint-react": "^5.0.0"
