+++ src/index.less
@@ -43,8 +43,8 @@
 
 @import "pages/detaljert-arbeidsforhold/DetaljertArbeidsforhold.less";
 @import "pages/endre-personopplysninger/EndreOpplysninger.less";
-@import "pages/digital-samhandling-offentlig-privat/DsopHistorikk.less";
-@import "pages/digital-samhandling-offentlig-privat/DsopDetaljer.less";
+@import "pages/digital-samhandling-offentlig-privat/historikk/DsopHistorikk.less";
+@import "pages/digital-samhandling-offentlig-privat/detaljer/DsopDetaljer.less";
 @import "pages/medlemskap-i-folketrygden/MedlHistorikk.less";
 @import "pages/institusjonsopphold/Inst.less";
 @import "pages/404/404.less";
