+++ src/pages/digital-samhandling-offentlig-privat/historikk/DsopHistorikk.tsx
@@ -1,4 +1,4 @@
-import React, { useState, Fragment, useEffect } from "react";
+import React, { Fragment, useEffect, useState } from "react";
 import { DsopInfo } from "types/dsop";
 import { Element } from "nav-frontend-typografi";
 import moment from "moment";
@@ -7,33 +7,13 @@ import Moment from "react-moment";
 import { NedChevron, OppChevron } from "nav-frontend-chevron";
 import { FormattedMessage } from "react-intl";
 import { Link, useLocation } from "react-router-dom";
-import PageContainer from "components/pagecontainer/PageContainer";
-import DSOPIkon from "assets/img/DSOP.svg";
-import WithDSOP from "./DsopFetch";
 import Lenke from "nav-frontend-lenker";
 
 interface Props {
   dsopInfo: DsopInfo;
 }
 
-/*
-  Hent data
-*/
-const DsopHistorikk = () => (
-  <PageContainer
-    tittelId={"dsop.tittel"}
-    icon={DSOPIkon}
-    backTo={"/#flere-opplysninger"}
-    brodsmulesti={[{ title: "dsop.tittel" }]}
-  >
-    <WithDSOP>{({ data }) => <Tabell dsopInfo={data} />}</WithDSOP>
-  </PageContainer>
-);
-
-/*
-  Visning
-*/
-const Tabell = (props: Props) => {
+export const DsopHistorikk = (props: Props) => {
   const location = useLocation();
   const { dsopInfo } = props;
